FIX - Delete active depends only for deleted issues.
git-svn-id: http://scarab.tigris.org/svn/scarab@11103 3c7165bb-d2ef-0310-b724-bafd9a86081f

+++ branches/release/1.0.22/src/java/org/tigris/scarab/om/Issue.java
@@ -3203,16 +3203,20 @@ public class Issue
             for (Iterator it = dependencies.iterator(); it.hasNext(); )
             {
                 Depend depend = (Depend)it.next();
-                ActivitySet deleteSet = 
-                    this.doDeleteDependency(activitySet, depend, user);
-                for (Iterator act = deleteSet.getActivityList().iterator(); act.hasNext(); )
-                {
-                    activitySet.addActivity((Activity)act.next());
-                }
-                NotificationManagerFactory.getInstance()
-                        .addActivityNotification(ActivityType.ISSUE_DELETED,
-                                activitySet, this, user);
+                if(!depend.getDeleted()){
+                	ActivitySet deleteSet = 
+                        this.doDeleteDependency(activitySet, depend, user);
+                    for (Iterator act = deleteSet.getActivityList().iterator(); act.hasNext(); )
+                    {
+                        activitySet.addActivity((Activity)act.next());
+                    }
+                }        
             }
+            
+            NotificationManagerFactory.getInstance()
+            .addActivityNotification(ActivityType.ISSUE_DELETED,
+                    activitySet, this, user);
+            
             save();
         }
     }
